rebase Jakub's changes on top of 1.11.0 release
3785003f 2020-04-07 | Allow explicit naming of method parameters in protocol to work around the breaking change in generation of protocols caused by removal of Paranamer dependency.[Jakub Drahos]
8baad6be 2020-04-06 | Create unions with null types in a way that they respect the default value type. Use parametrized types when available during parameter schema creation for ipc protocols. Using avro's lint before commit. [Jakub Drahos]
a5baf1a5 2020-04-03 | Allow nullable elements of arrays, maps and nullable record fields to be used as non null in array, map or record field default value declaration. In JacksonUtils#toJson write null node if the datum is null. In JacksonUtils#toObject handle nulls before handling union types so we are certain that the null was handled by the time we have to decide what type from the union to deserialize the json node to (we still don't know which union type the json node is, but if the first type is NULL second type of union should be used for deserialization) [Jakub Drahos]